New and updated routes.
These terms are needed for the creation of ECTO terms, in the context of ZAPP (the Zebrafish Atlas Toxicology Phenotype Project)

This PR add new exposure route, and updated some existing one to ensure that the definition and term names clearly refer to an exposure route.

Added "exposure regimen" branch. Though this new branch is similar to "temporal quality", it is not clear whether the "temporal quality" branch refer to an exposure regimen (e.g. definition are from PATO). I propose to review this at a later stage.